Refactor Partitioner::Maven
Move the code that is sharable to Go Partitioner to DependencyAnalysis
